COS-1779: Fix service line items optimistic query
