--- packages/apps/spaces/app/organization/[id]/src/components/Tabs/panels/AccountPanel/Contract/Services/modals/Service.dto.ts
@@ -1,80 +1,0 @@
-import { SelectOption } from '@shared/types/SelectOptions';
-import { billedTypeOptions } from '@organization/src/components/Tabs/panels/AccountPanel/utils';
-import {
-  BilledType,
-  ServiceLineItem,
-  ContractRenewalCycle,
-  ServiceLineItemInput,
-  ServiceLineItemUpdateInput,
-} from '@graphql/types';
-
-export interface ServiceForm {
-  name?: string;
-  price?: number;
-  reason?: string;
-  quantity?: number;
-  appSource?: string;
-  serviceStartedAt?: Date;
-  externalReference?: string;
-  isRetroactiveCorrection?: boolean;
-  renewalCycle?: ContractRenewalCycle;
-  billed?: SelectOption<BilledType> | null;
-}
-
-type ServiceItem = Omit<ServiceLineItem, 'id'>;
-
-export class ServiceDTO implements ServiceForm {
-  name?: string;
-  price?: number;
-  reason?: string;
-  quantity?: number;
-  appSource?: string;
-  billed?: SelectOption<BilledType> | null;
-  serviceStartedAt?: Date;
-  externalReference?: string;
-  renewalCycle?: ContractRenewalCycle;
-  isRetroactiveCorrection?: boolean;
-
-  constructor(data?: ServiceItem) {
-    this.quantity = data?.quantity;
-    this.name = data?.name ?? '';
-    this.price = data?.price;
-    this.appSource = data?.appSource;
-    this.billed =
-      billedTypeOptions.find((o) => o.value === data?.billed) ??
-      billedTypeOptions[2];
-    this.isRetroactiveCorrection = false;
-    this.reason = data?.comments || '';
-  }
-
-  static toForm(data?: ServiceItem): ServiceForm {
-    return new ServiceDTO(data);
-  }
-
-  static toPayload(
-    data: ServiceForm,
-    contractId: string,
-  ): ServiceLineItemInput {
-    return {
-      contractId,
-      name: data?.name,
-      price: data?.price,
-      quantity: data?.quantity,
-      appSource: data?.appSource,
-      billed: data?.billed?.value,
-    };
-  }
-  static toUpdatePayload(
-    data: ServiceForm,
-  ): Omit<ServiceLineItemUpdateInput, 'serviceLineItemId'> {
-    return {
-      name: data?.name,
-      price: data?.price,
-      comments: data?.reason,
-      quantity: data?.quantity,
-      appSource: data?.appSource,
-      billed: data?.billed?.value,
-      isRetroactiveCorrection: data?.isRetroactiveCorrection,
-    };
-  }
-}
